+++ tests.py
@@ -13,8 +13,9 @@ write("Testing with APSW file "+apsw.__file__+"\n")
 write("          APSW version "+apsw.apswversion()+"\n")
 write("    SQLite lib version "+apsw.sqlitelibversion()+"\n")
 write("SQLite headers version "+str(apsw.SQLITE_VERSION_NUMBER)+"\n")
+write("    Using amalgamation "+str(apsw.using_amalgamation)+"\n")
 
-if [int(x) for x in apsw.sqlitelibversion().split(".")]<[3,5,9]:
+if [int(x) for x in apsw.sqlitelibversion().split(".")]<[3,6,12]:
     write("You are using an earlier version of SQLite than recommended\n")
 
 sys.stdout.flush()